requirements.txt
sentence-transformers==5.3.0
chromadb==1.5.5
python-dotenv==1.2.2
requests==2.32.3=5